Add files via upload
This is a project that views the economic conditions for various countries, determines the best and worst region and country based on these economic freedoms, and concludes which group of economic freedoms was most related to a country’s index score using linear regression. Also, applied Techniques (k-means clustering) to group the countries into four subgroups and examine any similarities that existed within each group.